Ver 0.1.5
Web server now has basic convert functionality and can have a file uploaded and send via a POST request. The file is processed and saved server side, since this is all effectively plaintext it may be possible to just sent it to the browser in the post response and have it download.